requirements.txt
fastapi
uvicorn
motor
pdfminer.six
python-multipart
pydantic
tika
